requirements.txt
requests
psycopg2-binary
Flask
